refactor: CLI commands for seeding db
